refactor: move pre-aggregate logic from common to backend /ee
Moves all pre-aggregate business logic (parsing, matching, naming,
building, generation) from packages/common to packages/backend/src/ee.
Shared types (PreAggregateDef, PreAggregateMatchMiss) remain in common.

- Copy pre-aggregate modules to backend/src/ee/preAggregates/
- Update imports to use @lightdash/common for types, ee paths for logic
- Remove pre-aggregate module re-exports from common/src/index.ts
- Backend callers now import from ee/preAggregates/ instead of common
- Source files remain in common temporarily (used by translator tests)
