+++ packages/backend/src/ee/preAggregates/definition.ts
@@ -1,7 +1,9 @@
-import { type DbtPreAggregateDef } from '../types/dbt';
-import { ParseError } from '../types/errors';
-import { type PreAggregateDef } from '../types/preAggregate';
-import { TimeFrames } from '../types/timeFrames';
+import {
+    ParseError,
+    TimeFrames,
+    type DbtPreAggregateDef,
+    type PreAggregateDef,
+} from '@lightdash/common';
 
 const PRE_AGGREGATE_NAME_PATTERN = /^[a-zA-Z0-9_]+$/;
