+++ packages/backend/src/ee/preAggregates/metricRepresentation.ts
@@ -1,5 +1,4 @@
-import { MetricType } from '../types/field';
-import assertUnreachable from '../utils/assertUnreachable';
+import { assertUnreachable, MetricType } from '@lightdash/common';
 
 export type SupportedPreAggregateMetricType =
     | MetricType.SUM
